backend/requirements.txt
flask
torch
torchvision
Pillow
pandas
ultralytics